Cleaned up catchup structure;
