+++ packages/community-app/src/templates/Community/CommunityCatchup.tsx
@@ -1,0 +1,97 @@
+import { useCommunity } from '../../hooks/community/useCommunity';
+import { ErrorBoundary, LoadingBlock, t, useDotYouClient } from '@homebase-id/common-app';
+import { useParams } from 'react-router-dom';
+import { CommunityChannelCatchup } from '../../components/Community/catchup/CommunityChannelCatchup';
+import { CommunityChannelCatchupHeader } from '../../components/Community/catchup/CommunityChannelCatchup';
+import { useCommunityChannelsWithRecentMessages } from '../../hooks/community/channels/useCommunityChannelsWithRecentMessages';
+import { useCommunityMetadata } from '../../hooks/community/useCommunityMetadata';
+import { CommunityThread } from '../../components/Community/CommunityThread';
+
+export const CommunityChannelDetail = () => {
+  const { odinKey, communityKey: communityId, threadKey } = useParams();
+  const { data: community, isFetched } = useCommunity({ odinId: odinKey, communityId }).fetch;
+
+  const identity = useDotYouClient().getIdentity();
+  const { data: metadata } = useCommunityMetadata({
+    odinId: community?.fileMetadata.senderOdinId,
+    communityId: community?.fileMetadata?.appData?.uniqueId,
+  }).single;
+
+  const { data: channels } = useCommunityChannelsWithRecentMessages({
+    communityId: community?.fileMetadata?.appData?.uniqueId,
+  }).fetch;
+
+  const channelsToCatchup = channels?.filter((chnl) => {
+    if (!chnl.fileMetadata.appData.uniqueId) return false;
+    const lastReadTime =
+      metadata?.fileMetadata.appData.content.channelLastReadTime[
+        chnl.fileMetadata.appData.uniqueId
+      ];
+
+    return (
+      chnl.lastMessage?.fileMetadata.created &&
+      chnl.lastMessage.fileMetadata.created > (lastReadTime || 0) &&
+      !!chnl.lastMessage.fileMetadata.senderOdinId &&
+      chnl.lastMessage.fileMetadata.senderOdinId !== identity
+    );
+  });
+
+  if (!community && isFetched)
+    return (
+      <div className="flex h-full flex-grow flex-col items-center justify-center">
+        <p className="text-4xl">Homebase Community</p>
+      </div>
+    );
+
+  if (!community) {
+    return (
+      <div className="h-full w-full flex-grow bg-background">
+        <LoadingBlock className="h-16 w-full" />
+        <div className="mt-8 flex flex-col gap-4 px-5">
+          <LoadingBlock className="h-16 w-full" />
+          <LoadingBlock className="h-16 w-full" />
+          <LoadingBlock className="h-16 w-full" />
+          <LoadingBlock className="h-16 w-full" />
+          <LoadingBlock className="h-16 w-full" />
+        </div>
+      </div>
+    );
+  }
+
+  return (
+    <ErrorBoundary>
+      <div className="h-full w-full flex-grow bg-background">
+        <div className="relative flex h-full flex-row">
+          <div className="flex h-full flex-grow flex-col overflow-hidden">
+            <div className="flex h-full flex-grow flex-col">
+              <CommunityChannelCatchupHeader community={community} />
+              {!channelsToCatchup?.length ? (
+                <p className="m-auto text-lg">{t('All done!')} 🎉</p>
+              ) : (
+                <div className="flex h-20 flex-grow flex-col gap-3 overflow-auto p-3">
+                  {channelsToCatchup?.map((chnl) => (
+                    <CommunityChannelCatchup
+                      community={community}
+                      channel={chnl}
+                      key={chnl.fileId}
+                    />
+                  ))}
+                </div>
+              )}
+            </div>
+          </div>
+
+          {threadKey ? (
+            <ErrorBoundary>
+              <CommunityThread
+                community={community || undefined}
+                channel={undefined}
+                threadId={threadKey}
+              />
+            </ErrorBoundary>
+          ) : null}
+        </div>
+      </div>
+    </ErrorBoundary>
+  );
+};
